refactor(dev-agr): rename module package
