+++ algorithm/src/main/kotlin/io/github/wliamp/agr/data/Target.kt
@@ -1,4 +1,4 @@
-package io.github.wliamp.algorithm.data
+package io.github.wliamp.agr.data
 
 data class Target(
     val id: String,
