+++ algorithm/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.autoconfiguration.imports
@@ -1,1 +1,1 @@
-io.github.wliamp.algorithm.config.AlgorithmAutoConfig
+io.github.wliamp.agr.config.AlgorithmAutoConfig
