+++ algorithm/src/test/kotlin/io/github/wliamp/agr/TestClass.kt
@@ -1,4 +1,4 @@
-package io.github.wliamp.algorithm
+package io.github.wliamp.agr
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
